Update token mehtods
