+++ clubinho_preto/.env.example
@@ -14,8 +14,8 @@ ASAAS_URL=https://sandbox.asaas.com/api/v3/
 MELHORENVIO_URL=https://sandbox.melhorenvio.com.br/
 MELHORENVIO_CLIENT_ID=
 MELHORENVIO_SECRET=
-MELHORENVIO_REDIRECT_URL=
 MELHORENVIO_SHIPPING_FROM=
+BACKEND_BASE_URL=
 
 # Heorku only
 DISABLE_COLLECTSTATIC=1
