+++ clubinho_preto/settings.py
@@ -33,8 +33,10 @@ BASE_SUBSCRIPTION_VALUE = float(os.environ.get('BASE_SUBSCRIPTION_VALUE', defaul
 MELHORENVIO_URL = os.environ.get('MELHORENVIO_URL', default='https://www.melhorenvio.com.br/')
 MELHORENVIO_CLIENT_ID = os.environ.get('MELHORENVIO_CLIENT_ID', default='MELHORENVIO_CLIENT_ID')
 MELHORENVIO_SECRET = os.environ.get('MELHORENVIO_SECRET', default='MELHORENVIO_SECRET')
-MELHORENVIO_REDIRECT_URL = os.environ.get('MELHORENVIO_REDIRECT_URL', default='MELHORENVIO_REDIRECT_URL')
 MELHORENVIO_SHIPPING_FROM = os.environ.get('MELHORENVIO_SHIPPING_FROM', default='MELHORENVIO_SHIPPING_FROM')
+
+BACKEND_BASE_URL = os.environ.get('BACKEND_BASE_URL', default='BACKEND_BASE_URL')
+MELHORENVIO_REDIRECT_URL = f"{BACKEND_BASE_URL}get_token"
 
 # todo: restrict
 CORS_ALLOW_ALL_ORIGINS = True
