+++ melhor_envio/service.py
@@ -1,11 +1,58 @@
-from clubinho_preto.settings import (MELHORENVIO_CLIENT_ID,
-                                     MELHORENVIO_REDIRECT_URL,
-                                     MELHORENVIO_SECRET, MELHORENVIO_URL)
-import requests
-from django.core.cache import cache
 import json
 
+import requests
+from clubinho_preto.settings import (MELHORENVIO_CLIENT_ID, MELHORENVIO_SECRET,
+                                     MELHORENVIO_URL, MELHORENVIO_REDIRECT_URL)
+from django.core.cache import cache
+
+MELHORENVIO_CACHE_TIME = 2592000
+
+
+def save_token_to_cache(data):
+    cache.set("access_token", data.get("access_token"), MELHORENVIO_CACHE_TIME)
+    cache.set("refresh_token", data.get("refresh_token"), MELHORENVIO_CACHE_TIME*2)
+
+
 class MelhorEnvioService():
+
+    @staticmethod
+    def set_access_token(code):
+        response = requests.post(
+            f"{MELHORENVIO_URL}/oauth/token",
+            data={
+                "grant_type": "authorization_code",
+                "client_id": {MELHORENVIO_CLIENT_ID},
+                "client_secret": {MELHORENVIO_SECRET},
+                "redirect_uri": {MELHORENVIO_REDIRECT_URL},
+                "code": {code},
+            },
+        )
+        if response.ok:
+            # save tokens to cache
+            data = response.json()
+            save_token_to_cache(data)
+        return response
+
+    @staticmethod
+    def set_refresh_token(token=None):
+        if not token:
+            token = cache.get("refresh_token")
+
+        response = requests.post(
+            f"{MELHORENVIO_URL}/oauth/token",
+            data={
+                "grant_type": "refresh_token",
+                "client_id": {MELHORENVIO_CLIENT_ID},
+                "client_secret": {MELHORENVIO_SECRET},
+                "refresh_token": {token},
+            },
+        )
+
+        if response.ok:
+            # save tokens to cache
+            data = response.json()
+            save_token_to_cache(data)
+        return response
 
     @staticmethod
     def get_token():
@@ -50,7 +97,7 @@ class MelhorEnvioService():
             headers=headers,
             data=json.dumps(data)
         )
-        
+
         if response.ok:
             try:
                 return response.json()
