+++ melhor_envio/views.py
@@ -7,12 +7,7 @@ from django.http import HttpResponse
 from django.shortcuts import redirect
 from django.views import View
 
-MELHORENVIO_CACHE_TIME = 2592000
-
-
-def save_token_to_cache(data):
-    cache.set("access_token", data.get("access_token"), MELHORENVIO_CACHE_TIME)
-    cache.set("refresh_token", data.get("refresh_token"), MELHORENVIO_CACHE_TIME)
+from melhor_envio.service import MelhorEnvioService
 
 
 class AuthorizeApplicationView(View):
@@ -25,22 +20,7 @@ class AuthorizeApplicationView(View):
 class GetTokenView(View):
     def get(self, request):
         code = request.GET.get("code", None)
-
-        response = requests.post(
-            f"{MELHORENVIO_URL}/oauth/token",
-            data={
-                "grant_type": "authorization_code",
-                "client_id": {MELHORENVIO_CLIENT_ID},
-                "client_secret": {MELHORENVIO_SECRET},
-                "redirect_uri": {MELHORENVIO_REDIRECT_URL},
-                "code": {code},
-            },
-        )
-        if response.ok:
-            # save tokens to cache
-            data = response.json()
-            save_token_to_cache(data)
-
+        response = MelhorEnvioService.set_access_token(code=code)
         return HttpResponse(
             content=response.content, status=response.status_code, content_type=response.headers["Content-Type"]
         )
@@ -49,22 +29,7 @@ class GetTokenView(View):
 class RefreshTokenView(View):
     def get(self, request):
         refresh_token = request.GET.get("refresh_token", None)
-
-        response = requests.post(
-            f"{MELHORENVIO_URL}/oauth/token",
-            data={
-                "grant_type": "refresh_token",
-                "client_id": {MELHORENVIO_CLIENT_ID},
-                "client_secret": {MELHORENVIO_SECRET},
-                "refresh_token": {refresh_token},
-            },
-        )
-
-        if response.ok:
-            # save tokens to cache
-            data = response.json()
-            save_token_to_cache(data)
-
+        response = MelhorEnvioService.set_refresh_token(refresh_token)
         return HttpResponse(
             content=response.content, status=response.status_code, content_type=response.headers["Content-Type"]
         )
